Crate problem095

Problem 95 (Amicable chains)

Solution detail

This is another problem that brute-force is good enough for, assuming that you are computing the sum of divisors efficiently using prime factorisation.

Start by computing the sum of divisors of every number below the limit, and store them in an array (or store Option::None if the sum of divisors exceeds the limit).

Once the array has been pre-computed, just check every single number to see if it is part of a cycle, and keep track of the longest cycle seen and the smallest number in it.

Functions

answer
Solve the problem, returning the answer as a String

find_cycle_lengths
Find the cycle length of each starting number under the process n --> arr[n].

main

solve
Find the smallest member of the longest amicable chain with all members below the given limit.